Fix top links and buttons
HTML
The container that the right links and buttons are in exist so the
left links' container needs no extra margins.

The Google Sign-in button is actually a link that appears as a button,
so the button element was replaced with an anchor element.

The Google Apps button is actually a link with an ARIA role attribute
(button) along with other attributes, but the MDN documentation
suggests using a button instead, so that's what I used, along with the
ARIA label and expanded attributes.

CSS
Unused classes were removed and others were styled accordingly.